+++ GoDiveMVP/Views/Components/ExploreCatalogMapPinView.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+/// Explore catalog map push pin (red head).
+struct ExploreCatalogMapPinView: View {
+    var body: some View {
+        MapPushPinView(headColor: .red)
+    }
+}